// src/services/map/SanLuongKhachHang/models/donghokhachhang.model.ts
export interface DongHoKhachHang {
  OBJECTID?: number;
  IDMaDoiTuong?: string;
  TieuChuan?: number;
  NamSanXuat?: number;
  LoaiMatSo?: number;
  TinhTrang?: number;
  ApLuc?: number;
  DonViThiCong?: string;
  DonViThietKe?: string;
  MaDuong?: string;
  MaXa?: string;
  MaHuyen?: string;
  CoDongHo?: number;
  SoHopDong?: string;
  DBDongHoNuoc?: string;
  TenThueBao?: string;
  MucDichSuDung?: number;
  SoNguoiSuDung?: number;
  MaSoChi?: string;
  SoDienThoai?: string;
  SoNhaCu?: string;
  TinhTrangQLDHN?: number;
  TinhTrangThuNganDHN?: number;
  TinhTrangTuBoDHN?: number;
  CapChinhXac?: number;
  NgayLapDat?: number;
  NhaKiemDinh?: number;
  DinhMucSuDung?: number;
  LoaiDongHo?: number;
  MaVan?: string;
  TinhTrangVan?: number;
  GhiChuVan?: string;
  LoaiVan?: number;
  VatLieuVan?: number;
  HieuVan?: number;
  NamLapDatVan?: number;
  DoSauTyVan?: number;
  ChieuDongVan?: number;
  SVDongVan?: number;
  CoTyVan?: number;
  CoVan?: number;
  DonViQuanLy?: string;
  ViTriVan?: string;
  SoNhaMoi?: string;
  SoLanThay?: number;
  NgayThayMoi?: number;
  LyDoThayDHN?: string;
  GhiChu?: string;
  IDMaVan?: string;
  NguoiThiCong?: string;
  NguoiThietKe?: string;
  ChiSo?: number;
  Hieu?: string;
  ThoiGianDaSuDung?: number;
  ViTriLapDat?: number;
  TieuChuanVan?: number;
  NgayCapNhat?: number;
  NguoiCapNhat?: string;
  SoThanDongHo?: string;
  ID?: string;
  TenDuong?: string;
  NhanHieuDongHo?: string;
  SoChungNhanKiemDinh?: string;
  PhanLoaiChuSoHuu?: string;
  GhiChuSuaChuaChuKy?: string;
  ThoiGianKiemDinh?: number;
  CaoDoDatDongHo?: number;
  Layer?: string;
  SoSeri?: string;
  TenKhachHang_So?: string;
  NhanHieu_So?: string;
  HopBaoVeDongHo?: string;
  Enabled?: number;
  SoDB?: string;
}

export const DongHoKhachHangField = {
  OBJECTID: 'OBJECTID',
  IDMaDoiTuong: 'IDMaDoiTuong',
  TieuChuan: 'TieuChuan',
  NamSanXuat: 'NamSanXuat',
  LoaiMatSo: 'LoaiMatSo',
  TinhTrang: 'TinhTrang',
  ApLuc: 'ApLuc',
  DonViThiCong: 'DonViThiCong',
  DonViThietKe: 'DonViThietKe',
  MaDuong: 'MaDuong',
  MaXa: 'MaXa',
  MaHuyen: 'MaHuyen',
  CoDongHo: 'CoDongHo',
  SoHopDong: 'SoHopDong',
  DBDongHoNuoc: 'DBDongHoNuoc',
  TenThueBao: 'TenThueBao',
  MucDichSuDung: 'MucDichSuDung',
  SoNguoiSuDung: 'SoNguoiSuDung',
  MaSoChi: 'MaSoChi',
  SoDienThoai: 'SoDienThoai',
  SoNhaCu: 'SoNhaCu',
  TinhTrangQLDHN: 'TinhTrangQLDHN',
  TinhTrangThuNganDHN: 'TinhTrangThuNganDHN',
  TinhTrangTuBoDHN: 'TinhTrangTuBoDHN',
  CapChinhXac: 'CapChinhXac',
  NgayLapDat: 'NgayLapDat',
  NhaKiemDinh: 'NhaKiemDinh',
  DinhMucSuDung: 'DinhMucSuDung',
  LoaiDongHo: 'LoaiDongHo',
  MaVan: 'MaVan',
  TinhTrangVan: 'TinhTrangVan',
  GhiChuVan: 'GhiChuVan',
  LoaiVan: 'LoaiVan',
  VatLieuVan: 'VatLieuVan',
  HieuVan: 'HieuVan',
  NamLapDatVan: 'NamLapDatVan',
  DoSauTyVan: 'DoSauTyVan',
  ChieuDongVan: 'ChieuDongVan',
  SVDongVan: 'SVDongVan',
  CoTyVan: 'CoTyVan',
  CoVan: 'CoVan',
  DonViQuanLy: 'DonViQuanLy',
  ViTriVan: 'ViTriVan',
  SoNhaMoi: 'SoNhaMoi',
  SoLanThay: 'SoLanThay',
  NgayThayMoi: 'NgayThayMoi',
  LyDoThayDHN: 'LyDoThayDHN',
  GhiChu: 'GhiChu',
  IDMaVan: 'IDMaVan',
  NguoiThiCong: 'NguoiThiCong',
  NguoiThietKe: 'NguoiThietKe',
  ChiSo: 'ChiSo',
  Hieu: 'Hieu',
  ThoiGianDaSuDung: 'ThoiGianDaSuDung',
  ViTriLapDat: 'ViTriLapDat',
  TieuChuanVan: 'TieuChuanVan',
  NgayCapNhat: 'NgayCapNhat',
  NguoiCapNhat: 'NguoiCapNhat',
  SoThanDongHo: 'SoThanDongHo',
  ID: 'ID',
  TenDuong: 'TenDuong',
  NhanHieuDongHo: 'NhanHieuDongHo',
  SoChungNhanKiemDinh: 'SoChungNhanKiemDinh',
  PhanLoaiChuSoHuu: 'PhanLoaiChuSoHuu',
  GhiChuSuaChuaChuKy: 'GhiChuSuaChuaChuKy',
  ThoiGianKiemDinh: 'ThoiGianKiemDinh',
  CaoDoDatDongHo: 'CaoDoDatDongHo',
  Layer: 'Layer',
  SoSeri: 'SoSeri',
  TenKhachHang_So: 'TenKhachHang_So',
  NhanHieu_So: 'NhanHieu_So',
  HopBaoVeDongHo: 'HopBaoVeDongHo',
  Enabled: 'Enabled',
  SoDB: 'SoDB'
}